requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
filelock==3.18.0
fsspec==2025.5.1
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai-whisper @ git+https://github.com/openai/whisper.git@dd985ac4b90cafeef8712f2998d62c59c3e62d22
regex==2024.11.6
requests==2.32.4
setuptools==80.9.0
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.1
tqdm==4.67.1
typing_extensions==4.14.0
urllib3==2.5.0
